separation bloc RoomControls du fichier parent
